disttask: more log (#47575)
ref pingcap/tidb#46258
